feat: add cli script to run from command line
I generated this script by asking Claude 3.5 Sonnet to generate a CLI from the shiny app code.

I then of course reviewed and fixed the few bugs.

See README.md changes for an overview of how the CLI works and what it outputs.

In the future, it'd be easy to also optionally take a fasta file, run nextclade and use nextclade results as mutation input.